Refactored FieldFormatterCleanupsPanel to mvvm-pattern
