+++ src/main/java/org/jabref/gui/commonfxcontrols/FieldFormatterCleanupsPanel.fxml
@@ -1,0 +1,77 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.CheckBox?>
+<?import javafx.scene.control.ComboBox?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.Tooltip?>
+
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+
+<?import org.jabref.gui.icon.JabRefIconView?>
+
+<fx:root spacing="10.0" type="VBox"
+         xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1"
+         fx:controller="org.jabref.gui.commonfxcontrols.FieldFormatterCleanupsPanel">
+    <CheckBox fx:id="cleanupsEnabled" text="Enable"/>
+    <HBox spacing="4.0">
+        <VBox spacing="4.0" HBox.hgrow="ALWAYS">
+            <TableView fx:id="cleanupsList" minHeight="200.0" disable="${!cleanupsEnabled.selected}">
+                <columns>
+                    <TableColumn fx:id="fieldColumn" text="%Field name" reorderable="false"
+                                 minWidth="100.0" prefWidth="108.0"/>
+                    <TableColumn fx:id="formatterColumn" text="%Formatter name" reorderable="false"
+                                 minWidth="94.0" prefWidth="108.0"/>
+                    <TableColumn fx:id="actionsColumn" editable="false" reorderable="false" resizable="false" sortable="false"
+                                 maxWidth="30.0" minWidth="30.0" prefWidth="30.0"/>
+                </columns>
+                <columnResizePolicy>
+                    <TableView fx:constant="CONSTRAINED_RESIZE_POLICY"/>
+                </columnResizePolicy>
+            </TableView>
+            <HBox spacing="4.0">
+                <ComboBox fx:id="addableFields" editable="true" disable="${!cleanupsEnabled.selected}"
+                          HBox.hgrow="SOMETIMES" minWidth="100.0"/>
+                <ComboBox fx:id="addableFormatters" disable="${!cleanupsEnabled.selected}"
+                          HBox.hgrow="ALWAYS" minWidth="100.0" maxWidth="Infinity"/>
+            </HBox>
+        </VBox>
+        <VBox spacing="4.0">
+            <Button onAction="#resetToRecommended" disable="${!cleanupsEnabled.selected}"
+                    prefHeight="25.0" prefWidth="25.0" styleClass="icon-button,narrow">
+                <graphic>
+                    <JabRefIconView glyph="REFRESH"/>
+                </graphic>
+                <tooltip>
+                    <Tooltip text="%Reset to recommended"/>
+                </tooltip>
+            </Button>
+            <Button onAction="#clearAll" disable="${!cleanupsEnabled.selected}"
+                    prefHeight="25.0" prefWidth="25.0" styleClass="icon-button,narrow">
+                <graphic>
+                    <JabRefIconView glyph="CLEANUP_ENTRIES"/>
+                </graphic>
+                <tooltip>
+                    <Tooltip text="%Remove all"/>
+                </tooltip>
+            </Button>
+            <VBox VBox.vgrow="ALWAYS"/>
+            <Button onAction="#addCleanup" disable="${!cleanupsEnabled.selected}"
+                    prefHeight="25.0" prefWidth="25.0" styleClass="icon-button,narrow">
+                <graphic>
+                    <JabRefIconView glyph="ADD_NOBOX"/>
+                </graphic>
+                <tooltip>
+                    <Tooltip text="%Add"/>
+                </tooltip>
+            </Button>
+        </VBox>
+        <padding>
+            <Insets left="20.0"/>
+        </padding>
+    </HBox>
+</fx:root>
