+++ src/main/java/org/jabref/gui/libraryproperties/LibraryPropertiesDialogView.java
@@ -6,12 +6,12 @@ import java.util.Optional;
 
 import javax.inject.Inject;
 
+import javafx.collections.FXCollections;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.ButtonType;
 import javafx.scene.control.CheckBox;
 import javafx.scene.control.ComboBox;
-import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.VBox;
 
@@ -20,7 +20,9 @@ import org.jabref.gui.DialogService;
 import org.jabref.gui.commonfxcontrols.FieldFormatterCleanupsPanel;
 import org.jabref.gui.commonfxcontrols.SaveOrderConfigPanel;
 import org.jabref.gui.util.BaseDialog;
+import org.jabref.logic.cleanup.Cleanups;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.model.cleanup.FieldFormatterCleanups;
 import org.jabref.model.database.BibDatabaseMode;
 import org.jabref.model.metadata.MetaData;
 import org.jabref.model.metadata.SaveOrderConfig;
@@ -37,12 +39,13 @@ public class LibraryPropertiesDialogView extends BaseDialog<Void> {
     @FXML private TextField userSpecificFileDirectory;
     @FXML private TextField laTexFileDirectory;
     @FXML private CheckBox protect;
+    @FXML private FieldFormatterCleanupsPanel fieldFormatterCleanupsPanel;
     @Inject private PreferencesService preferencesService;
 
     private BasePanel panel;
     private LibraryPropertiesDialogViewModel viewModel;
     private final DialogService dialogService;
-    private FieldFormatterCleanupsPanel fieldFormatterCleanupsPanel;
+    private FieldFormatterCleanups oldFieldFormatterCleanups;
     private SaveOrderConfigPanel saveOrderConfigPanel;
     private SaveOrderConfig oldSaveOrderConfig;
 
@@ -85,11 +88,8 @@ public class LibraryPropertiesDialogView extends BaseDialog<Void> {
         oldSaveOrderConfig = storedSaveOrderConfig.orElseGet(preferencesService::loadExportSaveOrder);
 
         saveOrderConfigPanel.changeExportDescriptionToSave();
-        fieldFormatterCleanupsPanel = new FieldFormatterCleanupsPanel(Localization.lang("Enable save actions"));
-        Label saveActions = new Label(Localization.lang("Save actions"));
-        saveActions.getStyleClass().add("sectionHeader");
 
-        contentVbox.getChildren().addAll(saveOrderConfigPanel, saveActions, fieldFormatterCleanupsPanel);
+        contentVbox.getChildren().addAll(saveOrderConfigPanel);
 
         protect.selectedProperty().bindBidirectional(viewModel.libraryProtectedProperty());
 
@@ -97,7 +97,17 @@ public class LibraryPropertiesDialogView extends BaseDialog<Void> {
     }
 
     private void setValues() {
-        fieldFormatterCleanupsPanel.setValues(panel.getBibDatabaseContext().getMetaData());
+        Optional<FieldFormatterCleanups> saveActions = panel.getBibDatabaseContext().getMetaData().getSaveActions();
+        saveActions.ifPresentOrElse(value -> {
+            oldFieldFormatterCleanups = value;
+            fieldFormatterCleanupsPanel.cleanupsDisableProperty().setValue(!value.isEnabled());
+            fieldFormatterCleanupsPanel.cleanupsProperty().setValue(FXCollections.observableArrayList(value.getConfiguredActions()));
+        }, () -> {
+            oldFieldFormatterCleanups = Cleanups.DEFAULT_SAVE_ACTIONS;
+            fieldFormatterCleanupsPanel.cleanupsDisableProperty().setValue(!Cleanups.DEFAULT_SAVE_ACTIONS.isEnabled());
+            fieldFormatterCleanupsPanel.cleanupsProperty().setValue(FXCollections.observableArrayList(Cleanups.DEFAULT_SAVE_ACTIONS.getConfiguredActions()));
+        });
+
         saveOrderConfigPanel.setValues(oldSaveOrderConfig);
     }
 
@@ -167,19 +177,27 @@ public class LibraryPropertiesDialogView extends BaseDialog<Void> {
             }
         }
 
-        boolean saveActionsChanged = fieldFormatterCleanupsPanel.hasChanged();
-        if (saveActionsChanged) {
-            if (fieldFormatterCleanupsPanel.isDefaultSaveActions()) {
+        FieldFormatterCleanups fieldFormatterCleanups = new FieldFormatterCleanups(
+                !fieldFormatterCleanupsPanel.cleanupsDisableProperty().getValue(),
+                fieldFormatterCleanupsPanel.cleanupsProperty());
+
+        if (Cleanups.DEFAULT_SAVE_ACTIONS.equals(fieldFormatterCleanups)) {
+            metaData.clearSaveActions();
+        } else {
+            // if all actions have been removed, remove the save actions from the MetaData
+            if (fieldFormatterCleanups.getConfiguredActions().isEmpty()) {
                 metaData.clearSaveActions();
             } else {
-                fieldFormatterCleanupsPanel.storeSettings(metaData);
+                metaData.setSaveActions(fieldFormatterCleanups);
             }
         }
+
         boolean encodingChanged = !newEncoding.equals(oldEncoding);
 
         boolean changed = saveOrderConfigChanged || encodingChanged
-                          || viewModel.generalFileDirChanged() || viewModel.userFileDirChanged()
-                          || viewModel.protectedValueChanged() || saveActionsChanged || viewModel.laTexFileDirChanged();
+                || viewModel.generalFileDirChanged() || viewModel.userFileDirChanged()
+                || viewModel.protectedValueChanged() || viewModel.laTexFileDirChanged()
+                || !(oldFieldFormatterCleanups.equals(fieldFormatterCleanups));
         // ... if so, mark base changed. Prevent the Undo button from removing
         // change marking:
         if (changed) {
